+++ pkg/profiling/task/network/enums.go
@@ -17,6 +17,11 @@
 
 package network
 
+const (
+	unknown = "unknown"
+	http    = "http"
+)
+
 // ConnectionRole represents the role of the current process is the connection
 // whether it's a server or a client, if it's not trigger the connection/accept request, then it's unknown
 type ConnectionRole uint32
@@ -34,7 +39,7 @@ func (r ConnectionRole) String() string {
 	case ConnectionRoleServer:
 		return "server"
 	default:
-		return "unknown"
+		return unknown
 	}
 }
 
@@ -71,3 +76,51 @@ const (
 	SocketExceptionOperationRetransmit SocketExceptionOperationType = 1
 	SocketExceptionOperationDrop       SocketExceptionOperationType = 2
 )
+
+type ConnectionProtocol uint32
+
+const (
+	ConnectionProtocolUnknown ConnectionProtocol = 0
+	ConnectionProtocolHTTP    ConnectionProtocol = 1
+	ConnectionProtocolHTTP2   ConnectionProtocol = 2
+	ConnectionProtocolMySQL   ConnectionProtocol = 3
+	ConnectionProtocolCQL     ConnectionProtocol = 4
+	ConnectionProtocolPGSQL   ConnectionProtocol = 5
+	ConnectionProtocolDNS     ConnectionProtocol = 6
+	ConnectionProtocolRedis   ConnectionProtocol = 7
+	ConnectionProtocolNATS    ConnectionProtocol = 8
+	ConnectionProtocolMongo   ConnectionProtocol = 9
+	ConnectionProtocolKafka   ConnectionProtocol = 10
+	ConnectionProtocolMux     ConnectionProtocol = 11
+)
+
+func (c ConnectionProtocol) String() string {
+	switch c {
+	case ConnectionProtocolUnknown:
+		return unknown
+	case ConnectionProtocolHTTP:
+		return http
+	case ConnectionProtocolHTTP2:
+		return http
+	case ConnectionProtocolMySQL:
+		return "mysql"
+	case ConnectionProtocolCQL:
+		return "cql"
+	case ConnectionProtocolPGSQL:
+		return "pgsql"
+	case ConnectionProtocolDNS:
+		return "dns"
+	case ConnectionProtocolRedis:
+		return "redis"
+	case ConnectionProtocolNATS:
+		return "nats"
+	case ConnectionProtocolMongo:
+		return "mongo"
+	case ConnectionProtocolKafka:
+		return "kafka"
+	case ConnectionProtocolMux:
+		return "mutex"
+	default:
+		return unknown
+	}
+}